fixing pull request issues and started documentation
