+++ doc/docs/guides/search.md
@@ -13,3 +13,11 @@ The backend takes in a dictionary with the following fields
       'combine': whether the query is 'and' or 'or' in conjunction with other dictionaries
       'column': the model to be queried e.g. 'demographics'
 }
+
+
+### Custom Search Rules
+
+The Opal search allows you to write a custom rule, for example if you don't want to query by a subrecord.
+
+#### opal.core.search.SearchRule
+Is a [discoverable](../guides/discoverable.md)
