+++ opal/core/schemas.py
@@ -4,7 +4,7 @@ Utilities for dealing with Opal Schemas
 import itertools
 from opal.core.subrecords import subrecords
 from opal import models
-from opal.core.search.search_query import SearchQuery
+from opal.core.search.search_rule import SearchRule
 
 
 def serialize_model(model):
@@ -47,6 +47,6 @@ def list_records():
 
 
 def extract_schema():
-    custom_queries = [i().to_dict() for i in SearchQuery.list()]
+    custom_queries = [i().to_dict() for i in SearchRule.list()]
     schema = serialize_schema(itertools.chain([models.Tagging], subrecords()))
     return custom_queries + schema
